Change duplicated functions to an import export
