+++ routes/generatedMealsRoute.js
@@ -13,6 +13,8 @@ const { ObjectId } = require("mongodb");
 const Meal = require("../models/mealModel");
 // FavoriteMeal model
 const FavoriteMeal = require("../models/favMealModel");
+// Import function to query chatGPT
+const queryChatGPT = require('./queryChatGPT');
 
 // Available food tags
 const foodCategory = [
@@ -33,45 +35,6 @@ const foodCategory = [
   { name: "Seeds and Nuts" },
   { name: "Drinks,Alcohol, Beverages" },
 ];
-
-
-/**
- * Queries the OpenAI GPT-3.5 Turbo model for chat completions using the given prompt.
- * 
- * @async
- * @param {string} prompt the prompt to query the AI
- * @returns {Promise<string>} a Promise that resolves as a string with the response from the API.
- */
-async function queryChatGPT(prompt) {
-  const request = require("request");
-
-  const OPENAI_API_ENDPOINT = "https://api.openai.com/v1/chat/completions";
-
-  const options = {
-    url: OPENAI_API_ENDPOINT,
-    headers: {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${process.env.GPT_API_KEY}`,
-      "OpenAI-Organization": process.env.GPT_ORG_ID,
-    },
-    body: JSON.stringify({
-      model: "gpt-3.5-turbo",
-      messages: [{ role: "user", content: prompt }],
-      temperature: 0.7,
-    }),
-  };
-
-  return new Promise((resolve, reject) => {
-    request.post(options, (error, response, body) => {
-      if (error) {
-        console.error(error);
-        reject(error);
-      } else {
-        resolve(body);
-      }
-    });
-  });
-}
 
 
 /**
