+++ routes/queryChatGPT.js
@@ -1,0 +1,39 @@
+/**
+ * Queries the OpenAI GPT-3.5 Turbo model for chat completions using the given prompt.
+ * 
+ * @async
+ * @param {string} prompt the prompt to query the AI
+ * @returns {Promise<string>} a Promise that resolves as a string with the response from the API.
+ */
+async function queryChatGPT(prompt) {
+    const request = require("request");
+
+    const OPENAI_API_ENDPOINT = "https://api.openai.com/v1/chat/completions";
+
+    const options = {
+        url: OPENAI_API_ENDPOINT,
+        headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${process.env.GPT_API_KEY}`,
+            "OpenAI-Organization": process.env.GPT_ORG_ID,
+        },
+        body: JSON.stringify({
+            model: "gpt-3.5-turbo",
+            messages: [{ role: "user", content: prompt }],
+            temperature: 0.7,
+        }),
+    };
+
+    return new Promise((resolve, reject) => {
+        request.post(options, (error, response, body) => {
+            if (error) {
+                console.error(error);
+                reject(error);
+            } else {
+                resolve(body);
+            }
+        });
+    });
+}
+
+module.exports = queryChatGPT
